# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management API",
        "version": "1.0"
    },
    "paths": {
        "/api/events": {
            "get": {
                "summary": "List all upcoming events",
                "operationId": "6d8b6526460b262f52a64b734c86e021",
                "responses": {
                    "200": {
                        "description": "List of events"
                    }
                }
            },
            "post": {
                "summary": "Create a new event",
                "operationId": "26fc9153a1b63d68161c940df9168e0c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "start_time",
                                    "end_time",
                                    "max_capacity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "max_capacity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created"
                    }
                }
            }
        },
        "/api/events/{event_id}/register": {
            "post": {
                "summary": "Register attendee for event",
                "operationId": "f0f183a5dc3a6992ac0b908d25dc60c8",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered"
                    }
                }
            }
        },
        "/api/events/{event_id}/attendees": {
            "get": {
                "summary": "Get attendees for event with pagination",
                "operationId": "366f73eb201fdb65dcaa27e5d541d1ae",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated attendees"
                    }
                }
            }
        }
    }
}